Change directory name from baselines to algorithms
The name baselines makes conflicts with open-ai packages.
