+++ algorithms/bc/agent.py
@@ -18,9 +18,9 @@ import torch.optim as optim
 
 import wandb
 
-from baselines.ddpg.model import Actor, Critic
-from baselines.noise import OUNoise
-from baselines.replay_buffer import ReplayBuffer
+from algorithms.ddpg.model import Actor, Critic
+from algorithms.noise import OUNoise
+from algorithms.replay_buffer import ReplayBuffer
 
 
 # hyper parameters
@@ -34,7 +34,7 @@ hyper_params = {
         'EPISODE_NUM': 1500,
         'LAMBDA1': 1e-3,
         'LAMBDA2': 1.0,
-        'DEMO_PATH': "baselines/bc/demo_memory.pkl",
+        'DEMO_PATH': "data/lunarlander_continuous_demo.pkl",
 }
