+++ algorithms/dpg/agent.py
@@ -1,13 +1,14 @@
 # -*- coding: utf-8 -*-
-"""Reinforce agent for episodic tasks in OpenAI Gym.
+"""DPG agent for episodic tasks in OpenAI Gym.
 
 - Author: Curt Park
 - Contact: curt.park@medipixel.io
+- Paper: http://proceedings.mlr.press/v32/silver14.pdf
 """
 
 import os
 import git
-from collections import deque
+import numpy as np
 
 import torch
 import torch.nn.functional as F
@@ -15,20 +16,19 @@ import torch.optim as optim
 
 import wandb
 
-from baselines.reinforce.model import ActorCritic
+from algorithms.dpg.model import Actor, Critic
 
 
 # hyper parameters
 hyper_params = {
         'GAMMA': 0.99,
-        'STD': 1.0,
         'MAX_EPISODE_STEPS': 500,
         'EPISODE_NUM': 1500
 }
 
 
 class Agent(object):
-    """ReinforceAgent interacting with environment.
+    """ActorCritic interacting with environment.
 
     Args:
         env (gym.Env): openAI Gym environment with discrete action space
@@ -37,106 +37,83 @@ class Agent(object):
 
     Attributes:
         env (gym.Env): openAI Gym environment with discrete action space
-        model (nn.Module): policy gradient model to select actions
+        actor (nn.Module): actor model to select actions
+        critic (nn.Module): critic model to predict values
         args (dict): arguments including hyperparameters and training settings
-        optimizer (Optimizer): optimizer for training
-        log_prob_sequence (list): log probabailities of an episode
-        predicted_value_sequence (list): predicted values of an episode
-        reward_sequence (list): rewards of an episode to calculate returns
+        actor_optimizer (Optimizer): actor optimizer for training
+        critic_optimizer (Optimizer): critic optimizer for training
         device (torch.device): device selection (cpu / gpu)
 
     """
 
     def __init__(self, env, args, device):
         """Initialization."""
-        self.env = env
-        self.args = args
-        self.device = device
-        self.log_prob_sequence = []
-        self.predicted_value_sequence = []
-        self.reward_sequence = []
-
         # environment setup
+        self.env = env
         self.env._max_episode_steps = hyper_params['MAX_EPISODE_STEPS']
 
         # create a model
+        self.device = device
         state_dim = self.env.observation_space.shape[0]
         action_dim = self.env.action_space.shape[0]
         action_low = float(self.env.action_space.low[0])
         action_high = float(self.env.action_space.high[0])
-        self.model = ActorCritic(hyper_params['STD'], state_dim,
-                                 action_dim, action_low,
-                                 action_high).to(self.device)
+        self.actor = Actor(state_dim, action_dim, action_low,
+                           action_high, self.device).to(self.device)
+        self.critic = Critic(state_dim, action_dim,
+                             self.device).to(self.device)
 
         # create optimizer
-        self.optimizer = optim.Adam(self.model.parameters())
+        self.actor_optimizer = optim.Adam(self.actor.parameters())
+        self.critic_optimizer = optim.Adam(self.critic.parameters())
 
         # load stored parameters
         if args.load_from is not None and os.path.exists(args.load_from):
             self.load_params(args.load_from)
 
+        self.args = args
+
     def select_action(self, state):
         """Select an action from the input space."""
-        state = torch.from_numpy(state).float().to(self.device)
-        selected_action, predicted_value, dist = self.model(state)
+        selected_action = self.actor(state)
 
-        self.log_prob_sequence.append(dist.log_prob(selected_action).sum())
-        self.predicted_value_sequence.append(predicted_value)
-
-        return selected_action.detach().to('cpu').numpy()
+        return selected_action
 
     def step(self, action):
         """Take an action and return the response of the env."""
+        action = action.detach().to('cpu').numpy()
         next_state, reward, done, _ = self.env.step(action)
-
-        # store rewards to calculate return values
-        self.reward_sequence.append(reward)
 
         return next_state, reward, done
 
-    def update_model(self):
+    def update_model(self, experience):
         """Train the model after each episode."""
-        return_value = 0  # initial return value
-        return_sequence = deque()
+        state, action, reward, next_state, done = experience
 
-        # calculate return value at each step
-        for i in range(len(self.reward_sequence)-1, -1, -1):
-            return_value = self.reward_sequence[i] +\
-                           hyper_params['GAMMA'] *\
-                           return_value
-            return_sequence.appendleft(return_value)
+        # G_t   = r + gamma * v(s_{t+1})  if state != Terminal
+        #       = r                       otherwise
+        mask = 1 - done
+        value = self.critic(state, action)
+        next_action = self.actor(next_state)
+        next_value = self.critic(next_state, next_action).detach()
+        curr_return = reward + (hyper_params['GAMMA'] * next_value * mask)
+        curr_return = curr_return.to(self.device)
 
-        return_sequence = torch.tensor(return_sequence).to(self.device)
-        # standardize returns for better stability
-        return_sequence =\
-            (return_sequence - return_sequence.mean()) /\
-            (return_sequence.std() + 1e-7)
+        # train critic
+        critic_loss = F.mse_loss(value, curr_return)
+        self.critic_optimizer.zero_grad()
+        critic_loss.backward()
+        self.critic_optimizer.step()
 
-        # calculate loss at each step
-        loss_sequence = []
-        for log_prob,\
-            return_value,\
-            predicted_value in zip(self.log_prob_sequence,
-                                   return_sequence,
-                                   self.predicted_value_sequence):
-            delta = return_value - predicted_value.detach()
+        # train actor
+        action = self.actor(state)
+        actor_loss = -self.critic(state, action).mean()
+        self.actor_optimizer.zero_grad()
+        actor_loss.backward()
+        self.actor_optimizer.step()
 
-            policy_loss = -delta*log_prob
-            value_loss = F.smooth_l1_loss(predicted_value, return_value)
-
-            loss = (policy_loss + value_loss)
-            loss_sequence.append(loss)
-
-        # train
-        self.optimizer.zero_grad()
-        total_loss = torch.stack(loss_sequence).sum()
-        total_loss.backward()
-        self.optimizer.step()
-
-        # clear
-        self.log_prob_sequence.clear()
-        self.predicted_value_sequence.clear()
-        self.reward_sequence.clear()
+        # for logging
+        total_loss = critic_loss + actor_loss
 
         return total_loss.data
 
@@ -147,8 +124,12 @@ class Agent(object):
             return
 
         params = torch.load(path)
-        self.model.load_state_dict(params['model_state_dict'])
-        self.optimizer.load_state_dict(params['optim_state_dict'])
+        self.actor.load_state_dict(params['actor_state_dict'])
+        self.critic.load_state_dict(params['critic_state_dict'])
+        self.actor_optimizer.load_state_dict(
+                params['actor_optim_state_dict'])
+        self.critic_optimizer.load_state_dict(
+                params['critic_optim_state_dict'])
         print('[INFO] loaded the model and optimizer from', path)
 
     def save_params(self, n_episode):
@@ -157,13 +138,19 @@ class Agent(object):
             os.mkdir('./save')
 
         params = {
-                 'model_state_dict': self.model.state_dict(),
-                 'optim_state_dict': self.optimizer.state_dict()
+                 'actor_state_dict':
+                 self.actor.state_dict(),
+                 'critic_state_dict':
+                 self.critic.state_dict(),
+                 'actor_optim_state_dict':
+                 self.actor_optimizer.state_dict(),
+                 'critic_optim_state_dict':
+                 self.critic_optimizer.state_dict()
                  }
 
         repo = git.Repo(search_parent_directories=True)
         sha = repo.head.object.hexsha
-        path = os.path.join('./save/reinforce_' +
+        path = os.path.join('./save/dpg_continuous_' +
                             sha[:7] +
                             '_ep_' +
                             str(n_episode)+'.pt')
@@ -171,17 +158,19 @@ class Agent(object):
         print('[INFO] saved the model and optimizer to', path)
 
     def train(self):
-        """Run the agent."""
+        """Train the agent."""
         # logger
         if self.args.log:
             wandb.init()
             wandb.config.update(hyper_params)
-            wandb.watch(self.model, log='parameters')
+            wandb.watch(self.actor, log='parameters')
+            wandb.watch(self.critic, log='parameters')
 
         for i_episode in range(1, hyper_params['EPISODE_NUM']+1):
             state = self.env.reset()
             done = False
             score = 0
+            loss_episode = list()
 
             while not done:
                 if self.args.render and i_episode >= self.args.render_after:
@@ -189,16 +178,21 @@ class Agent(object):
 
                 action = self.select_action(state)
                 next_state, reward, done = self.step(action)
+                loss = self.update_model((state, action, reward,
+                                          next_state, done))
 
                 state = next_state
                 score += reward
 
+                loss_episode.append(loss)  # for logging
+
             else:
-                loss = self.update_model()
+                avg_loss = np.array(loss_episode).mean()
                 print('[INFO] episode %d\ttotal score: %d\tloss: %f'
-                      % (i_episode, score, loss))
+                      % (i_episode, score, avg_loss))
+
                 if self.args.log:
-                    wandb.log({'score': score, 'loss': loss})
+                    wandb.log({'score': score, 'avg_loss': avg_loss})
 
                 if i_episode % self.args.save_period == 0:
                     self.save_params(i_episode)
