+++ algorithms/noise.py
