+++ requirements.txt
@@ -1,55 +1,57 @@
 # Requirements automatically generated by pigar.
 # https://github.com/damnever/pigar
 
-# baselines/ac/agent.py: 9
-# baselines/bc/agent.py: 11
-# baselines/ddpg/agent.py: 10
-# baselines/dpg/agent.py: 10
-# baselines/ppo/agent.py: 10
-# baselines/reinforce/agent.py: 9
-# baselines/trpo/agent.py: 10
+# algorithms/ac/agent.py: 9
+# algorithms/bc/agent.py: 11
+# algorithms/ddpg/agent.py: 10
+# algorithms/dpg/agent.py: 10
+# algorithms/ppo/agent.py: 10
+# algorithms/reinforce/agent.py: 9
+# algorithms/trpo/agent.py: 11
 GitPython == 2.1.11
 
-# baselines/ppo/agent.py: 14
-# baselines/trpo/agent.py: 14
+# algorithms/ppo/agent.py: 14
+# run_lunarlander.py: 10
 gym == 0.10.9
 
-# baselines/ac/agent.py: 10
-# baselines/bc/agent.py: 13
-# baselines/ddpg/agent.py: 11
-# baselines/dpg/agent.py: 11
-# baselines/noise.py: 6
-# baselines/ppo/utils.py: 11
-# baselines/replay_buffer.py: 6
-# baselines/trpo/utils.py: 12
+# algorithms/ac/agent.py: 10
+# algorithms/bc/agent.py: 13
+# algorithms/ddpg/agent.py: 11
+# algorithms/dpg/agent.py: 11
+# algorithms/noise.py: 6
+# algorithms/ppo/utils.py: 11
+# algorithms/replay_buffer.py: 6
+# algorithms/trpo/utils.py: 12
 numpy == 1.15.4
 
-# baselines/trpo/agent.py: 17
+# algorithms/trpo/agent.py: 15
 scipy == 1.1.0
 
-# baselines/ac/agent.py: 12,13,14
-# baselines/ac/model.py: 11,12,13
-# baselines/bc/agent.py: 15,16,17
-# baselines/bc/model.py: 12,13
-# baselines/ddpg/agent.py: 13,14,15
-# baselines/ddpg/model.py: 12,13
-# baselines/dpg/agent.py: 13,14,15
-# baselines/dpg/model.py: 12,13,14
-# baselines/ppo/agent.py: 15,16,17,18
-# baselines/ppo/model.py: 12,13,14
-# baselines/ppo/utils.py: 12
-# baselines/reinforce/agent.py: 12,13,14
-# baselines/reinforce/model.py: 11,12,13
-# baselines/replay_buffer.py: 7
-# baselines/trpo/agent.py: 15,16
-# baselines/trpo/model.py: 12,13,14
-# baselines/trpo/utils.py: 13,14
+# algorithms/ac/agent.py: 12,13,14
+# algorithms/ac/model.py: 11,12,13
+# algorithms/bc/agent.py: 15,16,17
+# algorithms/bc/model.py: 12,13
+# algorithms/ddpg/agent.py: 13,14,15
+# algorithms/ddpg/model.py: 12,13
+# algorithms/dpg/agent.py: 13,14,15
+# algorithms/dpg/model.py: 12,13,14
+# algorithms/ppo/agent.py: 15,16,17,18
+# algorithms/ppo/model.py: 12,13,14
+# algorithms/ppo/utils.py: 12
+# algorithms/reinforce/agent.py: 12,13,14
+# algorithms/reinforce/model.py: 11,12,13
+# algorithms/replay_buffer.py: 7
+# algorithms/trpo/agent.py: 14
+# algorithms/trpo/model.py: 12,13,14
+# algorithms/trpo/utils.py: 13,14
+# run_lunarlander.py: 9
 torch == 0.4.1.post2
 
-# baselines/ac/agent.py: 16
-# baselines/bc/agent.py: 19
-# baselines/ddpg/agent.py: 17
-# baselines/dpg/agent.py: 17
-# baselines/ppo/agent.py: 23
-# baselines/reinforce/agent.py: 16
+# algorithms/ac/agent.py: 16
+# algorithms/bc/agent.py: 19
+# algorithms/ddpg/agent.py: 17
+# algorithms/dpg/agent.py: 17
+# algorithms/ppo/agent.py: 23
+# algorithms/reinforce/agent.py: 16
+# algorithms/trpo/agent.py: 20
 wandb == 0.6.29
